refactor(stock_balance): make use for frappe's `flt()` helper function
Make it a float value instead of integer

Also return the absolute value of the outgoing stock, negative doesn't make sense since its already __outgoing__

+++ accounting/accounting/doctype/stock_entry_table_item/stock_entry_table_item.py
@@ -18,7 +18,7 @@ class StockEntryTableItem(Document):
 		parent: DF.Data
 		parentfield: DF.Data
 		parenttype: DF.Data
-		quantity: DF.Int
+		quantity: DF.Float
 		rate: DF.Int
 		source_warehouse: DF.Link | None
 		target_warehouse: DF.Link | None
